Fix interval annotation types.
- Change the return type of `QuotientInterval::{min,max}` to `double`.
- Make the message in the exception thrown for a `@QuotientInterval` field
apply to the annotation type.

+++ source/soulboundarmory/module/config/QuotientInterval.java
@@ -8,7 +8,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
 public @interface QuotientInterval {
-	float min() default 0;
+	double min() default 0;
 
-	float max() default Float.MAX_VALUE;
+	double max() default Double.MAX_VALUE;
 }
